Small bugfix, added loading of model
