+++ vae_training/configs/CIFAR100_vae.yaml
@@ -14,6 +14,7 @@ dataset:
 
 
 vae:
+  load_from_checkpoint: null
   latent_dim: 32
   use_pretrained_cifar_enc: true
   im_channels: 3
